+++ libraries/pallas/src/pallas_read.cpp
@@ -201,7 +201,7 @@ Occurence* ThreadReader::getOccurence(pallas::Token id, size_t occurence_id) con
   return occurence;
 }
 
-AttributeList* ThreadReader::getEventAttributeList(Token event_id, int occurence_id) const {
+AttributeList* ThreadReader::getEventAttributeList(Token event_id, size_t occurence_id) const {
   auto* summary = getEventSummary(event_id);
   if (summary->attribute_buffer == nullptr)
     return nullptr;
@@ -217,7 +217,7 @@ AttributeList* ThreadReader::getEventAttributeList(Token event_id, int occurence
       return l;
     }
     if (l->index > occurence_id) {
-      pallas_error("Error fetching attribute %d. We went too far (cur position: %d) !\n", occurence_id, l->index);
+      pallas_error("Error fetching attribute %zu. We went too far (cur position: %d) !\n", occurence_id, l->index);
     }
   }
   return nullptr;
